Changed `modbus.c` to `modbus.cpp`.
The newer compiler used for the core I'm working with was unwilling to use `bool` in C files, and there wasn't a clean place to include `stdbool.h` without introducing other errors. Changing the file extension to `.cpp` (and mangling some variables named `template`, `and`, and `or`) reintroduced the expected semantics.